dropin/itk/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "itk"
version = "0.1.0"
description = "Offline-first integration test kit that renders sequence diagrams from logs/trace."
readme = "README_WORK_REPO.md"
requires-python = ">=3.10"
dependencies = [
  "boto3>=1.34.0",
  "botocore>=1.34.0",
  "pyyaml>=6.0.1",
  "jsonschema>=4.22.0",
  "rich>=13.7.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=4.1.0",
]

[project.scripts]
itk = "itk.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
